+++ linenoise.c
@@ -859,7 +859,7 @@ static void refreshLine(const char *prompt, struct current *current)
     /* Should intercept SIGWINCH. For now, just get the size every time */
     getWindowSize(current);
 
-    plen = strlen(prompt);
+    plen = (int)strlen(prompt);
     pchars = utf8_strwidth(prompt, utf8_strlen(prompt, plen));
 
     /* Scan the prompt for embedded ansi color control sequences and
@@ -954,8 +954,8 @@ static void set_current(struct current *current, const char *str)
 {
     strncpy(current->buf, str, current->bufmax);
     current->buf[current->bufmax - 1] = 0;
-    current->len = strlen(current->buf);
-    current->pos = current->chars = utf8_strlen(current->buf, current->len);
+    current->len = (int)strlen(current->buf);
+    current->pos = current->chars = (int)utf8_strlen(current->buf, current->len);
 }
 
 static int has_room(struct current *current, int bytes)
@@ -1300,7 +1300,7 @@ process_char:
                         if (rchars) {
                             int p = utf8_index(rbuf, --rchars);
                             rbuf[p] = 0;
-                            rlen = strlen(rbuf);
+                            rlen = (int)strlen(rbuf);
                         }
                         continue;
                     }
@@ -1353,7 +1353,7 @@ process_char:
                             }
                             /* Copy the matching line and set the cursor position */
                             set_current(current,history[searchpos]);
-                            current->pos = utf8_strlen(history[searchpos], p - history[searchpos]);
+                            current->pos = (int)utf8_strlen(history[searchpos], p - history[searchpos]);
                             break;
                         }
                     }
@@ -1516,7 +1516,7 @@ char *linenoise(const char *prompt)
         if (fgets(buf, sizeof(buf), stdin) == NULL) {
             return NULL;
         }
-        count = strlen(buf);
+        count = (int)strlen(buf);
         if (count && buf[count-1] == '\n') {
             count--;
             buf[count] = '\0';
